feat: move recenter button to map
