+++ shared-helpers/src/locales/es.json
@@ -1305,7 +1305,7 @@
   "t.range": "%{from} a %{to}",
   "t.readLess": "leer menos",
   "t.readMore": "leer más",
-  "t.recenterMap": "Vista de mapa más reciente",
+  "t.recenterMap": "Recentrar",
   "t.relationship": "Parentezco",
   "t.rent": "Alquiler",
   "t.review": "Revisión",
