+++ shared-helpers/src/locales/tl.json
@@ -1270,7 +1270,7 @@
   "t.range": "%{from} hanggang %{to}",
   "t.readLess": "basahin ang kaunti",
   "t.readMore": "basahin ang marami",
-  "t.recenterMap": "Recenter view ng mapa",
+  "t.recenterMap": "Recenter",
   "t.relationship": "Relasyon",
   "t.rent": "Renta",
   "t.review": "Review",
