+++ shared-helpers/src/locales/vi.json
@@ -1308,7 +1308,7 @@
   "t.range": "%{from} đến %{to}",
   "t.readLess": "đọc ít hơn",
   "t.readMore": "đọc nhiều hơn",
-  "t.recenterMap": "Chế độ xem bản đồ gần đây",
+  "t.recenterMap": "Gần đây hơn",
   "t.relationship": "Mối quan hệ",
   "t.rent": "Thuê",
   "t.review": "Xem xét",
