+++ shared-helpers/src/locales/zh.json
@@ -1308,7 +1308,7 @@
   "t.range": "由 %{from} 至 %{to}",
   "t.readLess": "閱讀較少內容",
   "t.readMore": "閱讀更多內容",
-  "t.recenterMap": "最近的地圖視圖",
+  "t.recenterMap": "近期中心",
   "t.relationship": "關係",
   "t.rent": "租金",
   "t.review": "審查",
