+++ sites/public/src/components/listings/ListingsList.tsx
@@ -1,13 +1,11 @@
 import * as React from "react"
-import { useMap } from "@vis.gl/react-google-maps"
 import { Listing, ListingMapMarker } from "@bloom-housing/shared-helpers/src/types/backend-swagger"
-import { Heading, Button } from "@bloom-housing/ui-seeds"
+import { Heading } from "@bloom-housing/ui-seeds"
 import { ZeroListingsItem } from "@bloom-housing/doorway-ui-components"
 import { LoadingOverlay, t, InfoCard, LinkButton } from "@bloom-housing/ui-components"
 import { getListings } from "../../lib/helpers"
 import { Pagination } from "./Pagination"
 import styles from "./ListingsCombined.module.scss"
-import { fitBounds } from "./MapClusterer"
 
 type ListingsListProps = {
   listings: Listing[]
@@ -16,12 +14,9 @@ type ListingsListProps = {
   onPageChange: (page: number) => void
   loading: boolean
   mapMarkers: ListingMapMarker[] | null
-  isDesktop: boolean
 }
 
 const ListingsList = (props: ListingsListProps) => {
-  const map = useMap()
-
   const moreMarkersOnMap = props.mapMarkers.length > 0
   const listingsDiv = (
     <div id="listingsList">
@@ -34,29 +29,7 @@ const ListingsList = (props: ListingsListProps) => {
         <ZeroListingsItem
           title={moreMarkersOnMap ? t("t.noVisibleListings") : t("t.noMatchingListings")}
           description={moreMarkersOnMap ? t("t.tryChangingArea") : t("t.tryRemovingFilters")}
-        >
-          {props.isDesktop && moreMarkersOnMap && (
-            <Button
-              onClick={() => {
-                if (map) {
-                  fitBounds(
-                    map,
-                    props.mapMarkers.map((marker, index) => {
-                      return {
-                        id: marker.id,
-                        key: index,
-                        coordinate: { lat: marker.lat, lng: marker.lng },
-                      }
-                    }),
-                    true
-                  )
-                }
-              }}
-            >
-              {t("t.recenterMap")}
-            </Button>
-          )}
-        </ZeroListingsItem>
+        />
       )}
     </div>
   )
