+++ sites/public/src/components/shared/MapControl.module.scss
@@ -7,8 +7,8 @@
   display: flex;
   flex-direction: column;
   position: absolute;
-  top: var(--bloom-s8);
-  left: var(--bloom-s8);
+  top: var(--seeds-s4);
+  left: var(--seeds-s4);
 
   .control-style {
     background-color: rgba(255, 255, 255, 0.85);
