+++ sites/public/src/components/shared/MapRecenter.module.scss
@@ -1,0 +1,5 @@
+.map-recenter {
+  position: absolute;
+  top: var(--seeds-s4);
+  right: var(--seeds-s4);
+}
