+++ sites/public/styles/overrides.scss
@@ -263,6 +263,7 @@
     &[data-variant="secondary"] {
       color: var(--doorway-color-blue-1300);
     }
+    line-height: normal;
   }
 
   .alert-box {
